# Add SubArray interface for better mem manegement