This PR tries to do better with allocations by:
 - Supporting `SubArray`, so that we can pass in views
 - Avoid allocating a big tuple of tuples for the unrolled combinations.

Closes #27. I'm not sure what's going on with allocations / performance:

```julia
julia> using BenchmarkTools, PlayingCards, PokerHandEvaluator

julia> using StatsBase: sample

julia> const fd = full_deck();

julia> sample_cards(N::Int) = sample_cards(Val(N))
sample_cards (generic function with 1 method)

julia> sample_cards(::Val{N}) where {N} = @view fd[sample(1:52, N; replace=false)]
sample_cards (generic function with 2 methods)

julia> @btime CompactHandEval(sample_cards(7))
  132.720 ms (424272 allocations: 26.70 MiB)
CompactHandEval{PokerHandEvaluator.HandTypes.TwoPair}(PokerHandEvaluator.HandTypes.TwoPair(), 2732)

julia> @btime CompactHandEval(sample_cards(7))
  21.955 ms (33420 allocations: 2.33 MiB)
CompactHandEval{PokerHandEvaluator.HandTypes.OnePair}(PokerHandEvaluator.HandTypes.OnePair(), 5539)

julia> @btime CompactHandEval(sample_cards(7))
  2.923 ms (3525 allocations: 203.67 KiB)
CompactHandEval{PokerHandEvaluator.HandTypes.TwoPair}(PokerHandEvaluator.HandTypes.TwoPair(), 3051)

julia> @btime CompactHandEval(sample_cards(7))
  59.600 μs (107 allocations: 2.28 KiB)
CompactHandEval{PokerHandEvaluator.HandTypes.TwoPair}(PokerHandEvaluator.HandTypes.TwoPair(), 2942)
```

Here is a sample of the size of allocations where we avoid creating the large tuple of combinations:
```julia
# main
@ballocated CompactHandEval(sample_cards(5)) # 288
@ballocated CompactHandEval(sample_cards(6)) # 976
@ballocated CompactHandEval(sample_cards(7)) # 9200
@ballocated FullHandEval(sample_cards(5)) # 816
@ballocated FullHandEval(sample_cards(6)) # 1920
@ballocated FullHandEval(sample_cards(7)) # 15680

# this PR
@ballocated CompactHandEval(sample_cards(5)) # 288
@ballocated CompactHandEval(sample_cards(6)) # 608
@ballocated CompactHandEval(sample_cards(7)) # 1728
@ballocated FullHandEval(sample_cards(5)) # 816
@ballocated FullHandEval(sample_cards(6)) # 944
@ballocated FullHandEval(sample_cards(7)) # 2384
```

It seems like we get roughly 3x slower what I would naively expect (~21*2 μs) at the end, but I'm not sure why it's changing so much each time. I tried first including the newly added file (`precompile`) which calls all the core `evaluate5` methods, but this didn't seem to help. Maybe the `precompile` file needs to call the top level function (e.g., `CompactHandEval`/`FulltHandEval`) instead of the core method (`evaluate5`)?

If it's any consolation, interpolating the entire view seems to measure great performance right from the start:
```julia
julia> using StatsBase, BenchmarkTools, PlayingCards, PokerHandEvaluator

julia> const fd = full_deck();

julia> @btime CompactHandEval($(@view fd[sample(1:52, 7; replace=false)]))
  11.400 μs (107 allocations: 1.69 KiB)
CompactHandEval{PokerHandEvaluator.HandTypes.HighCard}(PokerHandEvaluator.HandTypes.HighCard(), 6451)
```