@Moelf 2021-05-26
thanks for the super quick PR, I appreciate it!

first of all:
>but I'm not sure why it's changing so much each time. 

this is because `@btime` shows the least amount of time so it's luck-based, `@benchmark` would show you this.
```
julia> @benchmark CompactHandEval(sample_cards(7))
BenchmarkTools.Trial:
  memory estimate:  16.94 MiB
  allocs estimate:  268010
  --------------
  minimum time:     124.930 ms (0.00% GC)
  median time:      263.639 ms (3.89% GC)
  mean time:        269.180 ms (2.87% GC)
  maximum time:     410.603 ms (2.72% GC)
  --------------
  samples:          20
  evals/sample:     1
```

I also discovered that we sample directly by using `sample!`, which makes things significantly faster:
```
 julia> const fd = full_deck();

julia> const cards_buffer = Vector{Card}(undef, 7)
7-element Vector{Card}:

julia> @benchmark sample!($fd, cards_buffer; replace=false)
BenchmarkTools.Trial:
  memory estimate:  496 bytes
  allocs estimate:  1
  --------------
  minimum time:     192.065 ns (0.00% GC)
  median time:      204.055 ns (0.00% GC)
  mean time:        213.224 ns (1.62% GC)
  maximum time:     1.015 μs (71.41% GC)
  --------------
  samples:          10000
  evals/sample:     600
```
but yeah, the memory allocation is still staggering:
```
julia> @benchmark CompactHandEval(sample!($fd, $cards_buffer; replace=false))
BenchmarkTools.Trial:
  memory estimate:  1.29 MiB
  allocs estimate:  18700
  --------------
  minimum time:     20.113 ms (0.00% GC)
  median time:      164.844 ms (0.00% GC)
  mean time:        165.933 ms (2.72% GC)
  maximum time:     305.649 ms (4.25% GC)
  --------------
  samples:          31
  evals/sample:     1
```