@charleskawczynski 2021-05-26
> thanks for the super quick PR, I appreciate it!

My pleasure :slightly_smiling_face:, I'm almost at a point where I'm starting to think about performance in [TexasHoldem.jl](https://github.com/charleskawczynski/TexasHoldem.jl).

> I also discovered that we sample directly by using `sample!`, which makes things significantly faster:

Nice!

> but yeah, the memory allocation is still staggering:

True. A last resort might be to adjust PlayingCards.jl to be more like Stephan's Cards.jl, but I'd prefer to stick with the existing representation if we can make it work. I'm still seeing a decrease in allocations if `@benchmark` is called multiple times:

```julia
julia> @benchmark CompactHandEval(sample!($fd, $cards_buffer; replace=false))
BenchmarkTools.Trial:
  memory estimate:  2.63 MiB
  allocs estimate:  41205
  --------------
  minimum time:     45.936 ms (0.00% GC)
  median time:      118.131 ms (0.00% GC)
  mean time:        119.275 ms (0.47% GC)
  maximum time:     225.018 ms (2.30% GC)
  --------------
  samples:          42
  evals/sample:     1

julia> @benchmark CompactHandEval(sample!($fd, $cards_buffer; replace=false))
BenchmarkTools.Trial:
  memory estimate:  565.00 KiB
  allocs estimate:  8745
  --------------
  minimum time:     2.716 ms (0.00% GC)
  median time:      49.843 ms (0.00% GC)
  mean time:        54.908 ms (0.50% GC)
  maximum time:     161.007 ms (0.00% GC)
  --------------
  samples:          91
  evals/sample:     1

julia> @benchmark CompactHandEval(sample!($fd, $cards_buffer; replace=false))
BenchmarkTools.Trial:
  memory estimate:  2.11 KiB
  allocs estimate:  104
  --------------
  minimum time:     64.300 μs (0.00% GC)
  median time:      12.136 ms (0.00% GC)
  mean time:        19.564 ms (0.56% GC)
  maximum time:     97.030 ms (0.00% GC)
  --------------
  samples:          255
  evals/sample:     1
```

I don't really understand why, though.