@charleskawczynski 2021-06-16
Going to continue with this soon because TexasHoldem.jl is about ready for performance profiling and, IIRC, the first time I did it pointed to hand evaluations and creating the table (which allocates a full deck).