@charleskawczynski 2021-06-16
I think we need to fix precompilation before addressing this since timing seems to be pretty dramatically affecting the timings.