@charleskawczynski 2021-06-30
```julia
using PlayingCards
using PokerHandEvaluator
const fd = full_deck();
const cards_buffer = Vector{Card}(undef, 7)
@benchmark sample!($fd, cards_buffer; replace=false)
@benchmark CompactHandEval(sample!($fd, $cards_buffer; replace=false))
```