@charleskawczynski 2021-06-30
On `main` (without the view interface), we have:
```julia
julia> @benchmark CompactHandEval(Tuple(sample!($fd, $cards_buffer; replace=false)))
BechmarkTools.Trial: 10000 samples with 1 evaluations.
 Range (min … max):  11.208 μs … 243.429 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     26.165 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   27.308 μs ±   7.809 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

               ▃▅▇▇██▇▅▂                                        
  ▂▂▁▂▂▂▂▂▃▄▅▆███████████▆▅▄▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  11.2 μs         Histogram: frequency by time         61.7 μs <

 Memory estimate: 784 bytes, allocs estimate: 19.
```
Which seems to have slightly more allocations compared to the view interface, but the runtime is about the same. Maybe this whole issue was simply related to the PlayingCards memory layout (which is now much more compact) and type-stability (which, we're now type-stable using the new PlayingCards)? I guess it wouldn't hurt to add view support, but the tuple interface seems to work fine. @Moelf, are there are any other benefits to supporting views?