@charleskawczynski 2021-07-02
I think, between #33, #35, #36, we don't need this anymore. We can reopen if there's good reason, but performance seems to be _way_ better now, and I don't think the view will help much since each card has a _much_ smaller memory footprint.